feat: Create the Add route
