+++ src/components/Add/components/AddResultList/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./AddResultList";
