+++ src/components/Add/components/AddResults/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AddResults";
