+++ src/components/Add/components/index.ts
@@ -1,0 +1,2 @@
+export { default as AddResults } from "./AddResults";
+export { default as AddResultList } from "./AddResultList";
