--- src/components/Search/components/SearchContent/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./SearchContent";
