--- src/components/Search/components/index.ts
@@ -1,1 +1,0 @@
-export { default as SearchContent } from "./SearchContent";
